+++ Habiscus/Views/EditHabit/ActionSelectorView.swift
@@ -12,6 +12,7 @@ struct ActionSelectorView: View {
     @Environment(\.dismiss) private var dismiss
     @State private var editMode: EditMode = EditMode.active
     @Binding var actions: [Action]
+    @State var showActionInformation: ActionType? = nil
     
     var body: some View {
         NavigationStack {
@@ -19,11 +20,19 @@ struct ActionSelectorView: View {
                 Rectangle()
                     .fill(Color(UIColor.systemGroupedBackground))
                     .ignoresSafeArea()
+                if showActionInformation != nil {
+                    Rectangle().fill(.regularMaterial).ignoresSafeArea()
+                        .zIndex(1)
+                    ActionInfoView(information: $showActionInformation)
+                        .frame(maxWidth: 300, maxHeight: 400)
+                        .transition(.moveAndFade)
+                        .zIndex(2)
+                }
                 VStack(alignment: .leading) {
                         ScrollView(.horizontal) {
                             HStack {
                                 ForEach(ActionType.allCases) { action in
-                                    ActionListRowLarge(actions: $actions, action: action)
+                                    ActionListRowLarge(actions: $actions, action: action, information: $showActionInformation)
                                 }
                             }
                             .padding(.horizontal)
@@ -108,75 +117,6 @@ struct SelectedActionRow: View {
                 }
             }
         }
-    }
-}
-
-struct ActionListRowLarge: View {
-    @Environment(\.managedObjectContext) private var childContext
-    @Binding var actions: [Action]
-    let action: ActionType
-    
-    var body: some View {
-        ZStack(alignment: .topTrailing) {
-            VStack {
-                Button {
-                    withAnimation {
-                        addNewAction(type: action)
-                    }
-                } label: {
-                    Label("Add", systemImage: "plus.circle.fill")
-                        .font(.system(size: 20))
-                        .foregroundStyle(.white.opacity(0.9))
-                }
-                .labelStyle(.iconOnly)
-                Spacer()
-                Button {
-                    
-                } label: {
-                    Label("Add", systemImage: "info.circle.fill")
-                        .font(.system(size: 16))
-                        .foregroundStyle(.white.opacity(0.75))
-                }
-                .labelStyle(.iconOnly)
-            }
-            .padding(6)
-            ZStack(alignment: .bottomLeading) {
-                Image(systemName: action.symbol())
-                    .font(.system(size: 64))
-                    .foregroundStyle(.white.opacity(0.5))
-                    .offset(x: -20, y: 20)
-                VStack {
-                    Text(action.label())
-                        .font(.system(size: 20, weight: .semibold, design: .rounded))
-                        .foregroundStyle(.white)
-                    Spacer()
-                }
-                .frame(maxWidth: .infinity, alignment: .leading)
-            }
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-            .padding(8)
-            .clipShape(.rect(cornerRadius: 10))
-        }
-        .background(
-            RoundedRectangle(cornerRadius: 10)
-                .fill(action.color())
-                .shadow(color: .black.opacity(0.05), radius: 4, y: 2)
-                .shadow(color: action.color().opacity(0.25), radius: 4, y: 2)
-        )
-        .frame(width: 120, height: 120)
-        .padding(.top, 4)
-        .padding(.bottom, 12)
-    }
-    
-    func addNewAction(type: ActionType) {
-        let newAction = Action(context: childContext)
-        newAction.type = type.rawValue
-        if type == .timer {
-            newAction.number = 60.0
-        }
-        newAction.order = Int16(actions.count)
-        actions.append(newAction)
-        try? childContext.save()
     }
 }
